refactor flow create
